backend/requirements.txt
djangorestframework>=3.11.2
Django>=3.1.14
Pygments==2.7.4
Markdown==3.2.1
coreapi>=2.3.3
dj-database-url==0.5.0
gunicorn==20.0.4
whitenoise>=5.0.1
PyYAML==5.4
psycopg2>=2.8
mwclient==0.10.1
django-cors-headers>=3.13.0
django-oauth-toolkit>=2.2.0
pytz>=2022.7.1